% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/calculate_pve_of_signals.R
\name{calculate_pve_of_signals}
\alias{calculate_pve_of_signals}
\title{Calculates the percentage of variance explained for signals.}
\usage{
calculate_pve_of_signals(X, A, S, vcm)
}
\arguments{
\item{X}{Mixture matrix containing mixtures of signals in columns}

\item{A}{Mixing matrix}

\item{S}{Separated signals in columns of matrix}

\item{vcm}{vector containing column means of X}
}
\value{
List containing the percentage of varaince explained for each signal.
}
\description{
Calculates the percentage of variance explained for signals from mixtures
of signals.
}
\section{Warning}{
 There is currently no known formal proof
for this calculation.
}
\examples{
}
\seealso{
fastICA
}

